feat(player): 优化播放器断流重连和多人观看支持
- 播放器解复用错误时区分用户主动停止和自动断流，非用户停止则触发重连
- 重连后增加3秒等待时间确认视频帧到达，失败时安排下次重连尝试
- 实现多人观看共享流机制，复用现有活跃会话减少资源消耗
- 添加延迟关闭功能，最后一个观看者停止后等待60秒再真正断流
- 集成ZLM观看人数验证，在延迟关闭期间二次确认观看状态
- 修复前端重连状态管理，确保重连失败时正确显示错误信息
- 关闭HLS功能以优化性能，仅保留RTSP支持满足实时播放需求

+++ pkg/services/zlmediakit/types/media.go
@@ -16,20 +16,20 @@ type MediaOriginSock struct {
 }
 
 type MediaTrack struct {
-	Channels      int    `json:"channels"`
-	CodecId       int    `json:"codec_id"`
-	CodecIdName   string `json:"codec_id_name"`
-	CodecType     int    `json:"codec_type"`
-	Fps           int    `json:"fps"`
-	Frames        int64  `json:"frames"`
-	GopIntervalMs int64  `json:"gop_interval_ms"`
-	GopSize       int64  `json:"gop_size"`
-	KeyFrames     int64  `json:"key_frames"`
-	Ready         bool   `json:"ready"`
-	Height        int    `json:"height"`
-	Width         int    `json:"width"`
-	SampleBit     int    `json:"sample_bit"`
-	SampleRate    int    `json:"sample_rate"`
+	Channels      int     `json:"channels"`
+	CodecId       int     `json:"codec_id"`
+	CodecIdName   string  `json:"codec_id_name"`
+	CodecType     int     `json:"codec_type"`
+	Fps           float64 `json:"fps"`
+	Frames        int64   `json:"frames"`
+	GopIntervalMs int64   `json:"gop_interval_ms"`
+	GopSize       int64   `json:"gop_size"`
+	KeyFrames     int64   `json:"key_frames"`
+	Ready         bool    `json:"ready"`
+	Height        int     `json:"height"`
+	Width         int     `json:"width"`
+	SampleBit     int     `json:"sample_bit"`
+	SampleRate    int     `json:"sample_rate"`
 }
 
 type GetMediaListResp = Data[[]GetMediaListRespRaw]
